# Conflicts:
#	External/Plugins/ProjectManager/Helpers/LineEntryDialog.cs
#	FlashDevelop/Managers/ShortcutManager.cs
#	PluginCore/PluginCore/Interfaces.cs

+++ PluginCore/DockPanelSuite/Docking/DockPanel.AutoHideWindow.cs
@@ -298,7 +298,7 @@ namespace WeifenLuo.WinFormsUI.Docking
                 if (hovertime <= 0)
                     hovertime = 400;
 
-                m_timerMouseTrack.Interval = 2 * (int)hovertime;
+                m_timerMouseTrack.Interval = 2 * hovertime;
                 m_timerMouseTrack.Enabled = true;
             }
